+++ src/encryption.rs
@@ -1,11 +1,8 @@
 //! Module for encrypting and decrypting labels.
 
-use std::{
-    io::{Read as _, Write as _},
-    path::Path,
-};
+use std::{io::Write as _, path::Path};
 
-use age::secrecy::Secret;
+use age::secrecy::SecretString;
 use serde::{Deserialize, Serialize};
 
 use crate::{error::EncryptionError, Labels};
@@ -24,7 +21,7 @@ impl EncryptedLabels {
 
         let encrypted = {
             let encryptor =
-                age::Encryptor::with_user_passphrase(Secret::new(passphrase.to_owned()));
+                age::Encryptor::with_user_passphrase(SecretString::new(passphrase.into()));
 
             let mut encrypted = vec![];
             let mut writer = encryptor.wrap_output(&mut encrypted)?;
@@ -62,16 +59,9 @@ impl EncryptedLabels {
         let encrypted = &self.0;
 
         let decrypted = {
-            let decryptor = match age::Decryptor::new(&encrypted[..])? {
-                age::Decryptor::Passphrase(d) => d,
-                _ => unreachable!(),
-            };
-
-            let mut decrypted = vec![];
-            let mut reader = decryptor.decrypt(&Secret::new(passphrase.to_owned()), None)?;
-            reader.read_to_end(&mut decrypted)?;
-
-            decrypted
+            let secret_string = SecretString::new(passphrase.into());
+            let identity = age::scrypt::Identity::new(secret_string);
+            age::decrypt(&identity, encrypted)?
         };
 
         let labels_string = String::from_utf8(decrypted)?;