+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Update dependencies
+
 ## [0.1.1] - 2023-07-28
 
 - Documentation formatting